+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 #### NEXT
 * Add support for sACN protocol
 * Fix receiving higher ArtNet universes
-* Add  importing fixtures from MVR
+* Add import of fixtures from MVR
 * Support multipixel devices and color control per geometry
+* Show DMX Channel count in list of DMX Modes
+* Add Live DMX - a table with current DMX values
 * Add logging